Get rid of chpl_memalign (v. chpl_posix_memalign), mostly.
memalign() is problematic because you can't necessarily pass a pointer
returned by it back to free(), or even to realloc().  It's also not
POSIX compliant, and it's obsolete.  Here, mostly get rid of it.  Don't
call it anymore from anywhere in the Chapel repository, except for one
place where we're getting qthread stack pools which we should change in
their repo instead of ours.  We still provide chpl_memalign() for the
external interface, because that's public and we'll need to deprecate it
before removing it, but even that case is now implemented in terms of
chpl_posix_memalign().  Similarly, chpl-mem-replace-malloc.c which
replaces the standard libc memory allocation interface with functions
implemented in terms of Chapel memory allocator provides memalign()
still, but it's implemented in terms of posix_memalign().  And the layer
implementations for the runtime memory allocator no longer provide
chpl_memalign(), only chpl_posix_memalign().

This passed full testing, but on a MacBook Pro without LLVM, so it
didn't actually run the extern/ferguson/externblock/malloc_free_chpl
test that does a minimal check (chpl_malloc() and chpl_free()) of the
external Chapel memory allocator interface.  I'll arrange for that
testing next.