Don't hold the lock except when updating shared data.
This pushes the lock() and unlock() calls down into the functions that
add and remove memory tracking entries in the shared table.  Doing so
allows us to lock only the code that works with the table.  Most
importantly, it allows us to avoid holding the lock while allocating and
freeing tracking entries via the system allocator.