Add a test showing signal handling in Chapel (ish).
This program handles a signal and calls Chapel as part of that handling.
The handler itself isn't in Chapel, because it needs to take a C struct
whose definition can vary based on the platform.  But a Chapel function
is called from the handler.  The Chapel handler is quite limited in what
it can usefully do, however, since we have no idea what the program was
doing when the signal arrived, and only a very limited set of things are
safe to do in a signal handler anyway.