Skip maxTaskPar with synthetic NUMA domain count; output is weird.
Quoting from the comment in this new .skipif:
Skip the test if we're running with a synthetic NUMA domain count.
When we use that feature we don't end up with an entirely rational
topology and thus the program's output is somewhat bizarre.  Making
the .prediff match that would be hard.