If AM handler TX ctx uses a completion counter, count 'done' responses.
We started sending AM-related 'done' responses in a non-blocking way in
completion counter instead of a completion queue, we weren't counting
the 'done' indicator PUT as a transaction we had begun.  Failing to do
so meant that later when the tx counter incremented we had the wrong
expected value.  Here, count the PUT when we do it.