Refactor the provider selection code.
We can achieve Chapel MCM conformance using providers that support
either delivery-complete or a specific subset of the message orderings.
In the past we've preferred delivery-complete to message ordering, but
we have reason to believe we may be able to achieve better performance
with message ordering.  Here, refactor the provider selection logic to
make it a lot easier to go back and forth between these two, and perhaps
even others in the future.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>